pref: structure
