+++ src/markdown-it/index.ts
@@ -1,26 +1,19 @@
-import type MarkdownIt from "markdown-it"
+import MarkdownIt from "markdown-it"
+// @ts-ignore
+import mdKatex from "markdown-it-katex"
+import mdHighlight from "markdown-it-highlightjs"
+import mdKbd from "markdown-it-kbd"
+import preWrapperPlugin from "./preWrapper"
 
-export function preWrapperPlugin(md: MarkdownIt) {
-  const fence = md.renderer.rules.fence!
-  md.renderer.rules.fence = (...args) => {
-    const [tokens, idx] = args
-    const token = tokens[idx]
-    // remove title from info
-    token.info = token.info.replace(/\[.*\]/, "")
+const md = MarkdownIt({
+  linkify: true,
+  breaks: true
+})
+  .use(mdKatex)
+  .use(mdHighlight, {
+    inline: true
+  })
+  .use(mdKbd)
+  .use(preWrapperPlugin)
 
-    const lang = extractLang(token.info)
-    const rawCode = fence(...args)
-    return rawCode.replace(
-      "<pre>",
-      `<pre style="position: relative"><span class="lang">${lang}</span><button title="Copy Code" class="copy"></button>`
-    )
-  }
-}
-
-export function extractTitle(info: string) {
-  return info.match(/\[(.*)\]/)?.[1] || extractLang(info) || "txt"
-}
-
-const extractLang = (info: string) => {
-  return info.trim().replace(/:(no-)?line-numbers({| |$).*/, "")
-}
+export default md
